Reverted back to mesh-specific density field in two-phase model
Turns out it is more efficient to have a mesh-specific density field, because
then the expensive boundary update on the staggered rho field only needs to take
place once, when setting rho_. It is then used in the buoyancy function without
having to do a boundary update. The rhoMean calculation is now also cached.